Assez pour ce soir

+++ philo/utils.c
@@ -6,7 +6,7 @@
 /*   By: anadege <anadege@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/06 14:31:59 by anadege           #+#    #+#             */
-/*   Updated: 2021/08/10 21:00:29 by anadege          ###   ########.fr       */
+/*   Updated: 2021/08/10 21:18:55 by anadege          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ unsigned long	timestamp(void)
 	unsigned long	timestamp;
 
 	gettimeofday(&time, NULL);
-	if (g_launch == -1)
+	if ((int)g_launch == -1)
 	{
 		g_launch = 1000 * time.tv_sec + time.tv_usec / 1000;
 		return (0);
